requirements.txt
# Core async framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
uvloop==0.20.0



# LiveKit SDK
livekit
livekit-agents

# OpenAI
openai==1.45.0
elevenlabs==1.9.0

# VAD
torch==2.4.1
torchaudio==2.4.1
silero-vad==5.1.2

# Vector search
faiss-cpu==1.8.0
numpy==1.26.4

# Metrics
prometheus-client==0.21.0

# Audio processing
soundfile==0.12.1
librosa==0.10.2

# Utilities
python-dotenv==1.0.1
pydantic==2.9.0
httpx==0.27.2
aiohttp==3.9.5
